split application download tests in application and admin view. add test steps.
